feat: Optimize site performance and align brand profile
This commit introduces a comprehensive set of optimizations to improve the performance, accessibility, and maintainability of the wecanuseai.com portal.

Key changes include:
- Refactored the monolithic JavaScript file into smaller, lazy-loaded modules to reduce initial payload size.
- Implemented a build process with PostCSS and Terser to minify and purge unused CSS and JavaScript.
- Optimized the critical rendering path by inlining critical CSS, preloading key assets, and asynchronously loading non-critical stylesheets.
- Updated all site metadata (title, description, Open Graph tags) to align with the new brand focus on practical, privacy-conscious AI tools.
- Fixed all automated accessibility issues, achieving a perfect 100 score on Lighthouse for both mobile and desktop.
- Added a GitHub Actions workflow to automate the build and deployment process.
- Rewrote the README.md to provide clear instructions for development, building, and deployment.
