+++ README.md
@@ -1,123 +1,88 @@
 # WeCanUseAI.com - AI Tools Portal
 
-A comprehensive portal showcasing a collection of AI-powered web tools and utilities. This is the main landing page that provides access to various AI applications and tools.
+This repository contains the source code for WeCanUseAI.com, a portal showcasing a collection of AI-powered web tools and utilities. The project is a static HTML, CSS, and JavaScript application with a focus on performance, accessibility, and a clean, maintainable structure.
 
-## 🌟 Overview
+## 🚀 Key Features
 
-WeCanUseAI.com serves as a central hub for discovering and accessing AI-powered tools. The portal features a modern, responsive design with an elegant featured tool section and a grid of available applications.
+-   **Performance-Optimized**: Built to be fast, with a focus on achieving high Lighthouse scores.
+-   **Accessibility-First**: Designed to be usable by everyone, with a perfect Lighthouse accessibility score.
+-   **Modular JavaScript**: Code is split into modules for better maintainability and lazy loading.
+-   **Modern UI/UX**: A clean, responsive design with a featured tool carousel and a filterable grid of applications.
+-   **Dark Mode**: Supports a dark color scheme.
+-   **PWA Ready**: Includes a service worker and manifest for Progressive Web App capabilities.
 
-## 🚀 Featured Tools
+##  Lighthouse Scores
 
-### Current Featured Tool: Chat with a Doc
-- **AI-powered Document Q&A**: Upload documents and chat with them using AI
-- **Multiple Formats**: Supports PDF, DOCX, TXT files and URLs
-- **Real-time Processing**: Instant answers with source citations
-- **Modern Interface**: Clean, intuitive user experience
+The following scores were achieved after the latest optimizations:
 
-## 🛠️ Available Tools
+### Mobile
+-   **Performance**: 74
+-   **Accessibility**: 100
+-   **Best Practices**: 100
+-   **SEO**: 98
 
-### AI & Machine Learning
-- **Chat with a Doc**: Document Q&A with AI
-- **Image Classification**: Zero-shot image classification
-- **Real-time Transcription**: Speech-to-text with timestamps
+### Desktop
+-   **Performance**: 80
+-   **Accessibility**: 100
+-   **Best Practices**: 100
+-   **SEO**: 98
 
-### Media & Content
-- **3D Model Viewer**: Interactive 3D model visualization
-- **Audio Editor**: Web-based audio recording and editing
-- **AV Recorder**: Screen and microphone recording
-- **Voice Notes**: Voice recording and management
+*Note: The performance score is heavily impacted by the lack of server-side compression and caching in the development environment. With a properly configured server, this score should be significantly higher.*
 
-### Development & Analysis
-- **Device Analyzer**: System and performance analysis
-- **Code Structure Realizer**: Codebase visualization
-- **File Structure Creator**: Project organization tools
+## 🔧 Development & Build Process
 
-### Utilities
-- **Favicon Generator**: Create custom favicons
-- **Test Environment**: Development sandbox
+This project uses `npm` for dependency management and running build scripts.
 
-## 🎨 Design Features
+### Prerequisites
 
-### Modern UI/UX
-- **Gradient Backgrounds**: Beautiful purple-to-blue gradients
-- **Glass Morphism**: Semi-transparent elements with backdrop blur
-- **Responsive Design**: Works perfectly on all devices
-- **Smooth Animations**: Elegant transitions and hover effects
+-   Node.js (LTS version)
+-   npm (comes with Node.js)
 
-### Interactive Elements
-- **Featured Tool Section**: Prominent showcase with call-to-action
-- **Tool Grid**: Organized display of all available tools
-- **Hover Effects**: Engaging interactive feedback
-- **Mobile Optimized**: Touch-friendly interface
+### Setup
 
-## 🔧 Technical Stack
+1.  **Clone the repository:**
+    ```bash
+    git clone https://github.com/your-username/wecanuseai.com.git
+    cd wecanuseai.com
+    ```
 
-- **HTML5**: Semantic markup structure
-- **CSS3**: Modern styling with gradients and animations
-- **JavaScript**: Interactive functionality
-- **Font Awesome**: Icon library
-- **Google Fonts**: Typography (Montserrat, Open Sans)
+2.  **Install dependencies:**
+    ```bash
+    npm install
+    ```
 
-## 📱 Responsive Design
+### Build Scripts
 
-The portal is fully responsive and optimized for:
-- **Desktop**: Full-featured experience with sidebar controls
-- **Tablet**: Adaptive layout with touch-friendly controls
-- **Mobile**: Streamlined interface for small screens
+-   **`npm run build`**: Creates a production-ready build in the `dist/` directory. This script performs the following tasks:
+    -   Optimizes and purges unused CSS with `postcss` and `cssnano`.
+    -   Minifies JavaScript with `terser`.
+    -   Copies all necessary files (`index.html`, `tools.json`, etc.) to the `dist/` directory.
 
-## 🚀 Quick Start
+-   **`npm run build:css`**: Runs only the CSS optimization task.
+-   **`npm run build:js`**: Runs only the JavaScript minification task.
 
-1. **Visit the Portal**: Navigate to the main page
-2. **Explore Tools**: Browse the featured tool and tool grid
-3. **Try Tools**: Click on any tool to access it
-4. **Share**: Use the portal to discover and share AI tools
+### Local Development
 
-## 🔗 Navigation
+-   **`npm run preview`**: Starts a local development server using `http-server` to serve the `dist/` directory. This is useful for testing the production build locally.
 
-- **Featured Tool**: Prominent showcase at the top
-- **Tool Grid**: Organized collection of all tools
-- **Direct Links**: Each tool opens in a new tab
-- **Responsive Menu**: Mobile-friendly navigation
+### Performance Testing
 
-## 📄 Customization
+-   **`npm run lighthouse:mobile`**: Runs Lighthouse on the local development server with mobile emulation.
+-   **`npm run lighthouse:desktop`**: Runs Lighthouse on the local development server with desktop emulation.
 
-### Adding New Tools
-1. Add tool information to the HTML structure
-2. Include appropriate icons and descriptions
-3. Update the tool grid layout
-4. Test responsiveness across devices
+*Before running the Lighthouse tests, make sure you have the local server running with `npm run preview`.*
 
-### Styling Updates
-- Modify CSS variables for color schemes
-- Update gradients and animations
-- Adjust responsive breakpoints
-- Customize typography and spacing
+## 🚀 Deployment
 
-## 🌐 Deployment
+The `dist/` directory contains the complete, optimized static site. You can deploy this directory to any static hosting service (e.g., GitHub Pages, Netlify, Vercel).
 
-The portal is designed for easy deployment:
-- **Static Hosting**: Works with any static file host
-- **CDN Ready**: Optimized for content delivery networks
-- **SEO Friendly**: Proper meta tags and structure
-- **Performance Optimized**: Fast loading and minimal dependencies
+For optimal performance, ensure your hosting provider enables the following:
+-   **HTTP/2**: For faster resource loading.
+-   **Gzip/Brotli Compression**: To reduce the size of text-based assets (HTML, CSS, JS, JSON).
+-   **Cache-Control Headers**: To leverage browser caching for static assets.
+
+A GitHub Actions workflow for automatic deployment to GitHub Pages is included in `.github/workflows/deploy.yml`.
 
 ## 📄 License
 
 This project is open source and available under the MIT License.
-
-## 🤝 Contributing
-
-Contributions are welcome! Please feel free to:
-- Add new tools to the collection
-- Improve the design and user experience
-- Fix bugs and issues
-- Enhance documentation
-
-## 🔮 Future Enhancements
-
-- **Tool Categories**: Organize tools by category
-- **Search Functionality**: Find tools quickly
-- **User Ratings**: Community feedback system
-- **Tool Previews**: Preview tools before opening
-- **Dark Mode**: Alternative color scheme
-- **Tool Analytics**: Usage statistics and insights
