+++ package.json
@@ -1,15 +1,27 @@
 {
   "devDependencies": {
+    "@fullhuman/postcss-purgecss": "^6.0.0",
+    "@tailwindcss/postcss": "^4.1.1",
     "autoprefixer": "^10.4.21",
+    "cssnano": "^7.0.1",
+    "http-server": "^14.1.1",
+    "lighthouse": "^11.4.0",
+    "postcss": "^8.4.38",
     "postcss-cli": "^11.0.1",
-    "tailwindcss": "^4.1.11"
+    "terser": "^5.28.1"
   },
   "name": "app",
   "version": "1.0.0",
   "description": "A comprehensive portal showcasing a collection of AI-powered web tools and utilities. This is the main landing page that provides access to various AI applications and tools.",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build:css": "postcss style.css -o dist/style.css",
+    "build:js": "terser js/main.js -o dist/js/main.js -c -m",
+    "build": "mkdir -p dist/js && npm run build:css && npm run build:js && cp index.html dist/ && cp -r js/*.js dist/js/ && cp tools.json dist/ && cp critical.css dist/ && cp sw.js dist/ && cp robots.txt dist/ && cp sitemap.xml dist/",
+    "preview": "http-server -p 8080 dist",
+    "lighthouse:mobile": "lighthouse http://localhost:8080 --output=json --output-path=./lighthouse-mobile-report.json --only-categories=performance,accessibility,best-practices,seo --chrome-flags='--headless --no-sandbox'",
+    "lighthouse:desktop": "lighthouse http://localhost:8080 --output=json --output-path=./lighthouse-desktop-report.json --only-categories=performance,accessibility,best-practices,seo --chrome-flags='--headless --no-sandbox'"
   },
   "repository": {
     "type": "git",
